refactor(config-ui): move redux hook to hooks dir
