+++ config-ui/src/hooks/extend-redux.ts
@@ -17,7 +17,8 @@
  */
 
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from './store';
+
+import type { RootState, AppDispatch } from '@/app/store';
 
 type DispatchFunc = () => AppDispatch;
 export const useAppDispatch: DispatchFunc = useDispatch;
