+++ config-ui/src/hooks/index.ts
@@ -16,6 +16,7 @@
  *
  */
 
+export * from './extend-redux';
 export * from './use-auto-refresh';
 export * from './use-refresh-data';
 export * from './use-tips';
